requirements.txt
Django==2.2.*
django-extensions==2.2.*
djangorestframework==3.10.*
Pillow==6.1.0
pytz==2019.2

